bot/pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["urm8 <arctgisnil@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.20"
aiohttp = "^3.8.1"
python-dotenv = "^0.20.0"
furl = "^2.1.3"
pydantic = "^1.9.0"
beautifulsoup4 = "^4.11.1"
loguru = "^0.6.0"
aioredis = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
